+++ build_neighborhood.h
@@ -2,4 +2,4 @@
 
 double *build_neighborhood(Image *G, int L, int x, int y);
 
-int match_neighborhood(int *Na, int *Ns, int l);
+int match_neighborhood(double *Na, double *Ns, int l);
