+++ find_best_match.h
@@ -1,3 +1,3 @@
-#include <stdio.h>
+#include "image.h"
 
-int find_best_match(Image *Ga, Image *Gs, int L, int xs, int ys);
+double find_best_match(Image *Ga, Image *Gs, int L, int xs, int ys);
